requirements.txt
Pillow==9.4.0
PyYAML==6.0
requests==2.28.1
streamlit==1.12.0
streamlit_authenticator==0.2.1
streamlit_extras==0.2.6
